+++ src/lxc/storage/nbd.c
@@ -38,6 +38,7 @@
 #include "nbd.h"
 #include "storage.h"
 #include "storage_utils.h"
+#include "syscall_wrappers.h"
 #include "utils.h"
 
 #ifndef HAVE_STRLCPY
